Move userInteractions util to interactions folder
Renamed and relocated userInteractions.ts to src/utils/interactions/userInteractions.ts and updated import in AddStudent.ts to reflect the new path.
